+++ jetpack_vendor/i18n-map.php
@@ -50,7 +50,7 @@ return array(
     ),
     'jetpack-external-media' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-external-media',
-      'ver' => '0.1.0-alpha1738839433',
+      'ver' => '0.1.0-alpha1738894324',
     ),
     'jetpack-forms' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-forms',